+++ HISTORY.rst
@@ -28,6 +28,7 @@
 - Handle unicode filenames in staging folders.
 - Add staging file deletion permission.
 - New document_signature_view permission.
+- Add support for signing documents.
 
 2.0.2 (2016-02-09)
 ==================